--- a/.github/workflows/nodejs.yml
+++ b/.github/workflows/nodejs.yml
@@ -10,15 +10,17 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
-      - name: Use Node.js
-        uses: actions/setup-node@v2
+      - uses: actions/setup-node@v2
         with:
           node-version: '16'
           cache: 'npm'
           cache-dependency-path: src/package-lock.json
-      - name: Run test
+      - name: npm ci
         run: |
           cd src
           npm ci
+      - name: npm run build
+        run: |
+          cd src
           npm run build