NodeJS reduces build time by

- Activate the npm cache in action/setup-node
- Compile the AJV schema once and use it for both the positive and negative tests.

Performance improvement test on Mac mini:
make build: about 40 sec -> 30 sec
AJV part is about 27 sec -> 17 sec (This is where the 10 second time improvement comes from)
TV4 validator part is still about 13 sec.